turn blockheight into a custom hook
